+++ code/ryzom/client/src/http_client.cpp
@@ -79,7 +79,7 @@ bool CHttpClient::send(const std::string& buffer, bool verbose)
 	nlassert(_Sock.connected());
 
 	if(verbose) nldebug("Sending '%s' to '%s'", buffer.c_str(), _Sock.remoteAddr().asString().c_str());
-	uint32 size = buffer.size();
+	uint32 size = (uint32)buffer.size();
 	if(!buffer.empty())
 	{
 		if(_Sock.send((uint8 *)buffer.c_str(), size, false) != CSock::Ok)
