+++ code/ryzom/client/src/interface_v3/dbgroup_list_sheet.cpp
@@ -297,7 +297,7 @@ void CDBGroupListSheet::updateCoords ()
 		uint numValidSheets;
 		if (_Array)
 		{
-			numValidSheets = _SheetChildren.size();
+			numValidSheets = (uint)_SheetChildren.size();
 		}
 		else
 		{
